Added a method to grant permission based on trust.
This method grants permission to the given identity, creating a profile
for them if it does not already exist. This should only be called when
some mechanism has been used for establishing the validity of this
identity. Currently it is used by the web-services when provided with
trusted delegated authentication (verified by public key cryptography).


Former-commit-id: 0fffa47cf73c9616e500edb901aa6631c18360b1