Fix broken home page display on tomcat 8.
All along, calls to context.getRealPath() should have started with a / to reference the virtual file system.
However, only tomcat 8 enforces this
(http://stackoverflow.com/questions/25555541/getservletcontext-getrealpath-in-tomcat-8-returns-wrong-path)
We were using this call to detect whether we needed to load an extra css file for the page.
This change is fine with tomcat 7 as well.