Stop looking in sequence ontology for the track name, enabling us to use the JBrowse CanvasFeatures display type rather than the older HTMLFeatures (which doesn't support various track styles)
For the Gene track, for instance, we would get the correct type name (Gene) then promptly replace it with a lower-cased lookup from the SO (gene).
This would stop any track data showing up in JBrowse since it would be looking for a type with a different capitalization.
I don't see any obvious good reason for looking up the SO term name so I'm simply going to delete this step.