Only enable strand-specific region search if the new property genomicRegionSearch.enableStrandSpecificSearch == true
This is necessary because the new strand-specific checkbox in the region search page requires GenomicRegionSearchForm to accept a strandSpecific boolean property
But these properties are defined in a mine's resources/struts-config-form-model.xml file and existing mines will not have this entry.
If the property is not present then the region form fails to load properly.

To avoid issues with existing mines, this commit introduces a genomicRegionSearch.enableStrandSpecificSearch property to be set in <mine>/resources/web.properties.
If set to true, the mine must also have the following entry in its <mine>/resources/struts-config-form-model.xml

    <form-bean name="genomicRegionSearchForm" type="org.intermine.bio.web.struts.GenomicRegionSearchForm">
      ...
      <form-property name="strandSpecific" type="java.lang.Boolean"/>
    </form-bean>

genomicResearch.enableStrandSpecificSearch defaults to false