Allow the name attribute to be missing
Some OBO files have terms which are missing the name attribute. The code assumed that it exists, and the import therefore threw a fatal RunException when it hit a term which had a blank name attribute. This update simply tests for existence first, just as it already did for identifier (which, I thought, actually IS required in an OBO!).