Permanently disable CacheMapTest.test()
This is a terrible test because it relies on consistent memory pressure behaviour by the GC (which has changed when switching to the Gradle build system)
Testing properly requires a non-trivial mocking infrastructure.
But if it comes to that, don't do that but replace CacheMap, etc. with something battle-hardened from an external library - this is a common problem
that many others have tackled - WE DO NOT NEED AN INTERMINE SPECIFIC IMPLEMENTATION