Don't store term if name is missing, but keep on
The minor comment-out that was recently done to deal with missing names didn't cut it; the second set of code didn't have it commented out, so files like to.obo (Trait Ontology) still crashed. I think storing a term without a name is a Bad Idea, so I've simply changed the code to require that name exist to store a term, but not bail out if a term is missing the name. This successfully loads to.obo (after removing some include URLs at the top that no longer work correctly).