Update other-mines-links.js
I've commented out the DOMAIN_PARAMETER_NAME chunk because it was breaking the links to gene pages when present. For example, on LegumeMine, looking at a Cowpea gene, a homolog was listed on CowpeaMine, call it gene A; but when one clicked that link (which had the correct gene ID), CowpeaMine loaded the report page for a DIFFERENT gene. I say that's a bug. This fixes it.