At end of each PostProcessOperationsTask, close the underlying connection pool to stop gradle build failing.
FIXME: We have to do this to get Gradle working properly.  InterMine is only closing pools on
JVM exit but in Gradle everything is run in the same JVM, resulting in connection exhaustion.
InterMine should NEVER have relied on JVM exit, this whole thing needs to be fixed more
fundamentally