Fix NullPointerException in test
I am not sure what is going on here. It looks like Path.getEndFieldDescriptor() can return
null though so I think it makes sense to check for this. The tests pass but it looks like
both expected and actual results will go through the same code so I am not 100% confident
this is the right fix.