Standardise on javax.servlet-api v3.1.0
This was changed in an attempt to fix XML errors from Eclipse for web.xml.
Also involved removing strutstestcase which was pulling in v2.4. That will
be the next commit. In the end the errors turned out to be to do with the
use of non-standard metadata tag used for documentation.